[TBO] Move TBO + shared-experts-fusion e2e test to extra CI
The CI MLA proxy model lmsys/sglang-ci-dsv3-test has n_shared_experts != 1,
which cannot fuse the shared expert (do_load_weights asserts
num_fused_shared_experts == 1), so the test crashed at weight load on the
base-c deepep-4-gpu-h100 suite. Shared-experts fusion only supports exactly
one shared expert, so this needs a real DeepSeek-V3 (n_shared_experts=1).

Move TestTBOWithSharedExpertsFusion into its own file registered to the
label-gated extra CI (extra-b, deepep-8-gpu-h200), using deepseek-ai/
DeepSeek-V3-Base with tp8/dp8 (671B FP8 needs all 8 GPUs). Verified locally
on 8x H200: GSM8K 0.94. Revert test_deepep_small.py est_time 600 -> 478.
