+++ test/registered/ep/test_deepep_small.py
@@ -16,7 +16,7 @@ from sglang.test.test_utils import (
     popen_launch_server,
 )
 
-register_cuda_ci(est_time=600, stage="base-c", runner_config="deepep-4-gpu-h100")
+register_cuda_ci(est_time=478, stage="base-c", runner_config="deepep-4-gpu-h100")
 
 
 class TestPureDP(CustomTestCase):
@@ -225,58 +225,6 @@ class TestTBO(CustomTestCase):
                 "--enable-two-batch-overlap",
                 "--cuda-graph-max-bs",
                 "128",
-                "--max-running-requests",
-                "512",
-            ],
-            env={
-                **os.environ,
-                "SGLANG_TBO_DEBUG": "1",
-            },
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-    def test_gsm8k(self):
-        args = SimpleNamespace(
-            base_url=self.base_url,
-            model=self.model,
-            eval_name="gsm8k",
-            api="completion",
-            max_tokens=512,
-            num_examples=200,
-            num_threads=128,
-        )
-        metrics = run_eval(args)
-        print(metrics)
-
-        self.assertGreater(metrics["score"], 0.60)
-
-
-class TestTBOWithSharedExpertsFusion(CustomTestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.model = DEFAULT_MODEL_NAME_FOR_TEST_MLA
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=[
-                "--trust-remote-code",
-                "--tp",
-                "4",
-                "--enable-dp-attention",
-                "--dp",
-                "4",
-                "--moe-dense-tp-size",
-                "1",
-                "--moe-a2a-backend",
-                "deepep",
-                "--enable-two-batch-overlap",
-                "--enforce-shared-experts-fusion",
-                "--disable-cuda-graph",
                 "--max-running-requests",
                 "512",
             ],
